# New TableEditor, thread-safety, and table import