--- a/go/libraries/doltcore/doltdb/index_editor.go
+++ b/go/libraries/doltcore/doltdb/index_editor.go
@@ -18,6 +18,7 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"sync"
 
 	"github.com/liquidata-inc/dolt/go/libraries/doltcore/row"
 	"github.com/liquidata-inc/dolt/go/libraries/doltcore/schema"
@@ -28,27 +29,64 @@ import (
 )
 
 // IndexEditor takes in changes to an index map and returns the updated map if changes have been made.
-// This type is not thread-safe, and is intended for use in a single-threaded environment.
+//
+// This type is thread-safe, and may be used in a multi-threaded environment.
 type IndexEditor struct {
-	keyCount map[hash.Hash]int64
-	ed       *types.MapEditor
-	data     types.Map
-	idx      schema.Index
-	idxSch   schema.Schema // idx.Schema() builds the schema every call, so we cache it here
+	keyCount   map[hash.Hash]int64
+	ed         *types.MapEditor
+	data       types.Map
+	idx        schema.Index
+	idxSch     schema.Schema // idx.Schema() builds the schema every call, so we cache it here
+	needsFlush bool          // Whether the map editor has pending edits
+	updated    bool          // Whether the data has changed since the editor was created
+
+	// This mutex blocks on key count updates
+	keyMutex *sync.Mutex
+	// This mutex blocks on map edits
+	mapMutex *sync.Mutex
+	// This mutex ensures that Flush is only called once all current update operations have completed
+	flushMutex *sync.RWMutex
 }
 
 func NewIndexEditor(index schema.Index, indexData types.Map) *IndexEditor {
 	return &IndexEditor{
-		keyCount: make(map[hash.Hash]int64),
-		data:     indexData,
-		idx:      index,
-		idxSch:   index.Schema(),
+		keyCount:   make(map[hash.Hash]int64),
+		ed:         indexData.Edit(),
+		data:       indexData,
+		idx:        index,
+		idxSch:     index.Schema(),
+		needsFlush: false,
+		updated:    false,
+		keyMutex:   &sync.Mutex{},
+		mapMutex:   &sync.Mutex{},
+		flushMutex: &sync.RWMutex{},
 	}
 }
 
-// HasChanges returns whether the editor has any changes that need to be applied.
+// Flush applies all current edits to the underlying map.
+func (indexEd *IndexEditor) Flush(ctx context.Context) error {
+	indexEd.flushMutex.Lock()
+	defer indexEd.flushMutex.Unlock()
+
+	if indexEd.idx.IsUnique() {
+		for _, numOfKeys := range indexEd.keyCount {
+			if numOfKeys > 1 {
+				return fmt.Errorf("UNIQUE constraint violation on index: %s", indexEd.idx.Name())
+			}
+		}
+		indexEd.keyCount = make(map[hash.Hash]int64)
+	}
+	newIndexData, err := indexEd.ed.Map(ctx)
+	if err != nil {
+		return err
+	}
+	indexEd.reset(newIndexData)
+	return nil
+}
+
+// HasChanges returns whether the returned data would be different than the initial data.
 func (indexEd *IndexEditor) HasChanges() bool {
-	return indexEd.ed != nil
+	return indexEd.updated
 }
 
 // Index returns the index used for this editor.
@@ -56,23 +94,26 @@ func (indexEd *IndexEditor) Index() schema.Index {
 	return indexEd.idx
 }
 
-// Map applies all edits and returns a newly updated Map.
+// Map returns a Map based on the edits given, if any. If Flush() was not called prior, it will be called here.
 func (indexEd *IndexEditor) Map(ctx context.Context) (types.Map, error) {
-	if !indexEd.HasChanges() {
-		return indexEd.data, nil
-	}
-	if indexEd.Index().IsUnique() {
-		for _, numOfKeys := range indexEd.keyCount {
-			if numOfKeys > 1 {
-				return types.EmptyMap, fmt.Errorf("UNIQUE constraint violation on index: %s", indexEd.idx.Name())
-			}
+	indexEd.flushMutex.RLock() // if a Flush is ongoing then we need to wait
+	if indexEd.needsFlush {
+		indexEd.flushMutex.RUnlock() // Flush locks flushMutex, so we must unlock to prevent deadlock
+		err := indexEd.Flush(ctx)    // if this panics and is caught higher up then we are fine since we read unlocked
+		if err != nil {
+			return types.EmptyMap, err
 		}
+		indexEd.flushMutex.RLock() // we must read lock again since needsFlush may be false and we unlock in that case
 	}
-	return indexEd.ed.Map(ctx)
+	indexEd.flushMutex.RUnlock()
+	return indexEd.data, nil
 }
 
 // UpdateIndex updates the index map according to the given reduced index rows.
 func (indexEd *IndexEditor) UpdateIndex(ctx context.Context, originalIndexRow row.Row, updatedIndexRow row.Row) error {
+	indexEd.flushMutex.RLock()
+	defer indexEd.flushMutex.RUnlock()
+
 	if row.AreEqual(originalIndexRow, updatedIndexRow, indexEd.idxSch) {
 		return nil
 	}
@@ -81,9 +122,6 @@ func (indexEd *IndexEditor) UpdateIndex(ctx context.Context, originalIndexRow ro
 		indexKey, err := originalIndexRow.NomsMapKey(indexEd.idxSch).Value(ctx)
 		if err != nil {
 			return err
-		}
-		if indexEd.ed == nil {
-			indexEd.ed = indexEd.data.Edit()
 		}
 		if indexEd.idx.IsUnique() {
 			partialKey, err := originalIndexRow.ReduceToIndexPartialKey(indexEd.idx)
@@ -94,17 +132,20 @@ func (indexEd *IndexEditor) UpdateIndex(ctx context.Context, originalIndexRow ro
 			if err != nil {
 				return err
 			}
+			indexEd.keyMutex.Lock()
 			indexEd.keyCount[partialKeyHash]--
+			indexEd.keyMutex.Unlock()
 		}
+		indexEd.mapMutex.Lock()
 		indexEd.ed.Remove(indexKey)
+		indexEd.updated = true
+		indexEd.needsFlush = true
+		indexEd.mapMutex.Unlock()
 	}
 	if updatedIndexRow != nil {
 		indexKey, err := updatedIndexRow.NomsMapKey(indexEd.idxSch).Value(ctx)
 		if err != nil {
 			return err
-		}
-		if indexEd.ed == nil {
-			indexEd.ed = indexEd.data.Edit()
 		}
 		if indexEd.idx.IsUnique() {
 			partialKey, err := updatedIndexRow.ReduceToIndexPartialKey(indexEd.idx)
@@ -121,14 +162,29 @@ func (indexEd *IndexEditor) UpdateIndex(ctx context.Context, originalIndexRow ro
 				}}})
 			_, err = mapIter.ReadRow(ctx)
 			if err == nil { // row exists
+				indexEd.keyMutex.Lock()
 				indexEd.keyCount[partialKeyHash]++
+				indexEd.keyMutex.Unlock()
 			} else if err != io.EOF {
 				return err
 			}
+			indexEd.keyMutex.Lock()
 			indexEd.keyCount[partialKeyHash]++
+			indexEd.keyMutex.Unlock()
 		}
+		indexEd.mapMutex.Lock()
 		indexEd.ed.Set(indexKey, updatedIndexRow.NomsMapValue(indexEd.idxSch))
+		indexEd.updated = true
+		indexEd.needsFlush = true
+		indexEd.mapMutex.Unlock()
 	}
 
 	return nil
+}
+
+func (indexEd *IndexEditor) reset(indexData types.Map) {
+	indexEd.keyCount = make(map[hash.Hash]int64)
+	indexEd.ed = indexData.Edit()
+	indexEd.data = indexData
+	indexEd.needsFlush = false
 }