… go/go.sum
… go/libraries/doltcore/doltdb/index_editor_test.go
… go/libraries/doltcore/doltdb/table_editor_test.go
… go/libraries/doltcore/doltdb/table_test.go
… go/libraries/doltcore/mvdata/data_loc_test.go
… go/libraries/doltcore/mvdata/table_data_loc.go
… go/libraries/doltcore/sqle/table_editor.go
… go/libraries/doltcore/sqle/table_editor_test.go
… go/libraries/doltcore/sqle/tables.go