So I created a new `TableEditor` under `doltdb`, similar to the `IndexEditor`. I renamed the old editor to `sqlTableEditor`, and it's now using `TableEditor` underneath. It's just to comply with the SQL interfaces.

Perhaps one of the most striking changes are the synchronization changes. The motivation was from a comment from @zachmu , which implied that some of the write paths are parallelized for performance, and the old editor was not thread-safe by any means. In order to support these new write paths, it was necessary to make `TableEditor` thread-safe. Because of this, `IndexEditor` also received the same changes. I verified that there was a lack of data races using `go test -race`, and also ran the concurrency tests with thousands more iterations and edits (they're lower now so that we're not spending an hour in Jenkins). Just to verify that my tests also worked, I made temporary intentional changes that would allow for data races, and the tests caught them, and `go test -race` also caught them.

In addition to the above changes, `dolt table import` and `dolt table cp` both use the `TableEditor` now. Due to the concerns mentioned in the previous paragraph, I decided to benchmark a few metrics to get an idea of the performance impact of these two major changes.

To measure just the impact of the threading changes, I imported 1,000,000 rows through SQL and timed the results:
```
Before: 67.074s
After:  67.363s
```
The result is close enough that it's within the margin of error, so it's safe to say that new `TableEditor` is just as performant as the old implementation.

The `table import` is a different story though. This time, I imported a 10,000,000 row `.psv`, with the results being:
```
Before: 53.521s
After:  107.816s
```
It takes roughly twice as long now, which is slower, but not tragically so. Primarily, the difference in speed is that the old code just wrote directly to the edit accumulator, where as the `TableEditor` does a ton more bookkeeping, being a generalized editor.