@Hydrocharged 2020-05-28
Even if the pipeline code does not currently use multiple threads for writing to the `TableEditor`, we'll still have to make it thread-safe at some point in the future. I'd say these should have been thread-safe from the beginning, and not having them so does limit their potential usage.

I kept the lock implementations simple for the sake of clarity as well. We can go back over them later whenever do a performance overhaul of the editors. For `TableEditor`, it's all just `Lock()/defer Unlock()`, so it's not hard to trace.

For `IndexEditor`, I'm not using the `defer Unlock()` pattern so much, as the `Lock()/Unlock()` just surrounds portions of code that are 1-3 lines long and don't have reasons to panic, so it's also straightforward in its implementation.