@zachmu 2020-05-28
I haven't looked at this yet, but wanted to address the performance regression.

A 50% performance penalty in CSV import is maybe barely permissible, but should be avoided if at all possible, even if it makes the code more complicated. This is the front door for most customers, especially those with large pre-existing data sets. It needs to be as fast as possible. I think 10M rows in 2 min is probably acceptable, but obv faster is better. It's worth doing some work profiling this to see where the time is going for CSV, and which parts could be avoided. If there are easy wins to be had here, let's take them.

I agree substantially about the thread safety concern (better to assume it will be used in a non-thread-safe way). Probably worth benchmarking with and without the locks to see if it costs us performance, though.