@Hydrocharged 2020-05-28
I'll profile and see what I can come up with. I tried some simple multithreaded changes in the import code and only got around 2 seconds of speed increase for a lot more complexity, so I reverted that and shipped what is here.

> Probably worth benchmarking with and without the locks to see if it costs us performance, though.

Darn it @zachmu it's right there in my initial post lol, I'll just copy it here too:

> To measure just the impact of the threading changes, I imported 1,000,000 rows through SQL and timed the results:
> ```
> Before: 67.074s
> After:  67.363s
> ```
> The result is close enough that it's within the margin of error, so it's safe to say that new > `TableEditor` is just as performant as the old implementation.