@zachmu 2020-05-29
Brian, this is about a 2x performance regression for import. I want you to see if you have any ideas on how to speed it up. Daylon is in the middle of profiling it, so he might have some low hanging fruit too.