@Hydrocharged 2020-05-29
So from profiling...well I didn't find anything that can be considered "low hanging fruit". It looks like a significant chunk of time spent is on garbage collection, and I can't find anything quick and simple that can reduce that. Besides that though, the load is fairly spread out.

Performance of this path is something that I can really dig into and try to optimize after foreign keys are in. Since everything will be using this new path, it would lead to speed increases across the entire product.